+++ src/App.js
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <>
+      <h1>hello</h1>
+    </>
+  );
+}
+
+export default App;
